+++ packages/subgraph/tests/dao/dao_v1_0_0.test.ts
@@ -60,11 +60,6 @@ import {
   ExtendedNativeTransfer,
 } from '../helpers/extended-schema';
 import {
-  createDummyActions,
-  createERC1155TokenCalls,
-  createTokenCalls,
-} from '../utils';
-import {
   getBalanceOf,
   createNewExecutedEvent,
   createDaoEntityState,
@@ -81,6 +76,9 @@ import {
   generateTokenIdBalanceEntityId,
   generateTransactionActionsProposalEntityId,
   generateTransferEntityId,
+  createDummyAction,
+  createERC20TokenCalls,
+  createERC1155TokenCalls,
 } from '@aragon/osx-commons-subgraph';
 import {Address, Bytes, BigInt, ethereum} from '@graphprotocol/graph-ts';
 import {
@@ -120,13 +118,13 @@ function createExecutedEvent(
       isDynamic
     );
 
-    let action = createDummyActions(
+    let action = createDummyAction(
       DAO_TOKEN_ADDRESS,
       '0',
       functionData.toHexString()
     );
 
-    actions.push(action[0]);
+    actions.push(action);
   }
 
   if (execResults.length == 0) {
@@ -260,12 +258,16 @@ describe('handleNativeTokenDeposited', () => {
 describe('handleDeposited: ', () => {
   beforeAll(() => {
     daoTokenContract = new ExtendedERC20Contract().withDefaultValues();
-    let totalSupply = '10';
-    daoTokenContract.mockCall_createTokenCalls(totalSupply);
+
+    createERC20TokenCalls(
+      DAO_TOKEN_ADDRESS,
+      ERC20_AMOUNT_FULL,
+      'DAO Token',
+      'DAOT',
+      '6'
+    );
     daoTokenContract.mockCall_balanceOf(DAO_ADDRESS, ERC20_AMOUNT_HALF);
     daoTokenContract.mockCall_balanceOf(DAO_TOKEN_ADDRESS, ERC20_AMOUNT_HALF);
-
-    createTokenCalls(DAO_TOKEN_ADDRESS, 'DAO Token', 'DAOT', null, null);
 
     getSupportsInterface(DAO_TOKEN_ADDRESS, ERC165_INTERFACE_ID, true);
     getSupportsInterface(
@@ -993,7 +995,7 @@ describe('handleExecuted', () => {
 
   describe('ERC20 action', () => {
     beforeAll(() => {
-      createTokenCalls(DAO_TOKEN_ADDRESS, 'name', 'symbol', '6', '10');
+      createERC20TokenCalls(DAO_TOKEN_ADDRESS, '10', 'name', 'symbol', '6');
       getBalanceOf(DAO_TOKEN_ADDRESS, DAO_ADDRESS, ERC20_AMOUNT_HALF);
       getBalanceOf(DAO_TOKEN_ADDRESS, DAO_TOKEN_ADDRESS, ERC20_AMOUNT_HALF);
 
@@ -1247,7 +1249,7 @@ describe('handleExecuted', () => {
 
   describe('ERC721 action', () => {
     beforeAll(() => {
-      createTokenCalls(DAO_TOKEN_ADDRESS, 'name', 'symbol', null, null);
+      createERC20TokenCalls(DAO_TOKEN_ADDRESS, '10', 'name', 'symbol');
 
       getSupportsInterface(DAO_TOKEN_ADDRESS, '0x01ffc9a7', true);
       getSupportsInterface(DAO_TOKEN_ADDRESS, '80ac58cd', true);