+++ packages/subgraph/tests/dao/dao_v1_3_0.test.ts
@@ -36,11 +36,6 @@ import {
   ExtendedERC1155Transfer,
 } from '../helpers/extended-schema';
 import {
-  createDummyActions,
-  createERC1155TokenCalls,
-  createTokenCalls,
-} from '../utils';
-import {
   createNewExecutedEvent,
   encodeWithFunctionSelector,
   getBalanceOf,
@@ -54,6 +49,9 @@ import {
   generateTokenIdBalanceEntityId,
   generateTransactionActionsProposalEntityId,
   generateTransferEntityId,
+  createDummyAction,
+  createERC20TokenCalls,
+  createERC1155TokenCalls,
 } from '@aragon/osx-commons-subgraph';
 import {ethereum, Bytes, Address, BigInt} from '@graphprotocol/graph-ts';
 import {
@@ -250,7 +248,7 @@ describe('handleExecuted', () => {
 
   describe('ERC20 action', () => {
     beforeAll(() => {
-      createTokenCalls(DAO_TOKEN_ADDRESS, 'name', 'symbol', '6', '10');
+      createERC20TokenCalls(DAO_TOKEN_ADDRESS, '10', 'name', 'symbol', '6');
       getBalanceOf(DAO_TOKEN_ADDRESS, DAO_ADDRESS, ERC20_AMOUNT_HALF);
       getBalanceOf(DAO_TOKEN_ADDRESS, DAO_TOKEN_ADDRESS, ERC20_AMOUNT_HALF);
 
@@ -267,7 +265,7 @@ describe('handleExecuted', () => {
 
     describe('ERC20 transfer action', () => {
       beforeAll(() => {
-        createTokenCalls(DAO_TOKEN_ADDRESS, 'name', 'symbol', null, null);
+        createERC20TokenCalls(DAO_TOKEN_ADDRESS, '10', 'name', 'symbol');
 
         getSupportsInterface(DAO_TOKEN_ADDRESS, ERC165_INTERFACE_ID, true);
         getSupportsInterface(
@@ -510,7 +508,7 @@ describe('handleExecuted', () => {
 
   describe('ERC721 action', () => {
     beforeAll(() => {
-      createTokenCalls(DAO_TOKEN_ADDRESS, 'name', 'symbol', null, null);
+      createERC20TokenCalls(DAO_TOKEN_ADDRESS, '10', 'name', 'symbol');
 
       getSupportsInterface(DAO_TOKEN_ADDRESS, '0x01ffc9a7', true);
       getSupportsInterface(DAO_TOKEN_ADDRESS, '80ac58cd', true);
@@ -1048,13 +1046,13 @@ function createExecutedEvent(
       isDynamic
     );
 
-    let action = createDummyActions(
+    let action = createDummyAction(
       DAO_TOKEN_ADDRESS,
       '0',
       functionData.toHexString()
     );
 
-    actions.push(action[0]);
+    actions.push(action);
   }
 
   if (execResults.length == 0) {