+++ packages/subgraph/tests/utils.ts
@@ -1,4 +1,4 @@
-import {Address, BigInt, Bytes, ethereum} from '@graphprotocol/graph-ts';
+import {Address, BigInt, ethereum} from '@graphprotocol/graph-ts';
 import {createMockedFunction} from 'matchstick-as/assembly/index';
 
 export function createMockGetter(
@@ -14,78 +14,6 @@ export function createMockGetter(
   )
     .withArgs([])
     .returns(returns);
-}
-
-export function createERC1155TokenCalls(
-  contractAddress: string,
-  tokenId: string,
-  uri: string
-): void {
-  createMockedFunction(
-    Address.fromString(contractAddress),
-    'uri',
-    'uri(uint256):(string)'
-  )
-    .withArgs([ethereum.Value.fromUnsignedBigInt(BigInt.fromString(tokenId))])
-    .returns([ethereum.Value.fromString(uri)]);
-}
-
-export function createTokenCalls(
-  contractAddress: string,
-  name: string,
-  symbol: string,
-  decimals: string | null,
-  totalSupply: string | null
-): void {
-  createMockGetter(contractAddress, 'name', 'name():(string)', [
-    ethereum.Value.fromString(name),
-  ]);
-
-  createMockGetter(contractAddress, 'symbol', 'symbol():(string)', [
-    ethereum.Value.fromString(symbol),
-  ]);
-
-  if (decimals) {
-    createMockGetter(contractAddress, 'decimals', 'decimals():(uint8)', [
-      ethereum.Value.fromUnsignedBigInt(BigInt.fromString(decimals)),
-    ]);
-  }
-
-  if (totalSupply) {
-    createMockGetter(
-      contractAddress,
-      'totalSupply',
-      'totalSupply():(uint256)',
-      [ethereum.Value.fromUnsignedBigInt(BigInt.fromString(totalSupply))]
-    );
-  }
-}
-
-export function createWrappedTokenCalls(
-  contractAddress: string,
-  name: string,
-  symbol: string,
-  underlyingTokenAddress: string,
-  totalSupply: string | null
-): void {
-  createTokenCalls(contractAddress, name, symbol, '18', totalSupply);
-  createMockGetter(contractAddress, 'underlying', 'underlying():(address)', [
-    ethereum.Value.fromAddress(Address.fromString(underlyingTokenAddress)),
-  ]);
-}
-
-export function createDummyActions(
-  address: string,
-  value: string,
-  data: string
-): ethereum.Tuple[] {
-  let tuple = new ethereum.Tuple();
-
-  tuple.push(ethereum.Value.fromAddress(Address.fromString(address)));
-  tuple.push(ethereum.Value.fromSignedBigInt(BigInt.fromString(value)));
-  tuple.push(ethereum.Value.fromBytes(Bytes.fromHexString(data) as Bytes));
-
-  return [tuple];
 }
 
 export function createGetProposalCall(